LPS-116840 Extract showSidebar logic and actions to util
